Using the createTask method
Is there a difference? As far as I can see no, but it works like this,
and I have no idea (pun intended) why. And I don't really want to
spend the rest of life figuring out through the mist of whosever logic
this belongs to. Some things are better left as they are...